Fixed Fire Red crash
Gen III Bulbfinder was assuming a start location from Emerald's Pokemon Data address, causing it to fail to find it when acting on unmodified Fire Red. This caused the program to incorrectly modify other data, leading to a crash when entering the southern Viridian Forest gatehouse (and probably other locations, but this is the first encountered in regular play).